create tests
create tests

+++ src/main/java/com/evatool/variants/domain/events/VariantEventListener.java
@@ -6,6 +6,8 @@ import com.evatool.global.event.analysis.AnalysisDeletedEvent;
 import com.evatool.global.event.requirements.RequirementCreatedEvent;
 import com.evatool.global.event.requirements.RequirementDeletedEvent;
 import com.evatool.global.event.requirements.RequirementUpdatedEvent;
+import com.evatool.variants.common.error.exceptions.EventEntityAlreadyExistsException;
+import com.evatool.variants.common.error.exceptions.EventEntityDoesNotExistException;
 import com.evatool.variants.common.error.exceptions.IllegalEventPayloadException;
 import com.evatool.variants.domain.entities.VariantsAnalysis;
 import com.evatool.variants.domain.entities.VariantsRequirements;
@@ -38,6 +40,11 @@ public class VariantEventListener {
     public void analyseCreated(AnalysisCreatedEvent event){
         logger.info("analyse created event");
         if(logger.isDebugEnabled())logger.debug(String.format(DEBUGFORMAT,event.getClass(), event.getJsonPayload()));
+
+        if (variantsAnalysisRepository.existsById(VariantsAnalysis.fromJson(event.getJsonPayload()).getAnalysisId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
+
         try {
             VariantsAnalysis variantsAnalysis = gson.fromJson(event.getJsonPayload(), VariantsAnalysis.class);
             variantsAnalysisRepository.save(variantsAnalysis);
@@ -52,6 +59,11 @@ public class VariantEventListener {
     public void analyseDeleted(AnalysisDeletedEvent event){
         logger.info("analyse created event");
         if(logger.isDebugEnabled())logger.debug(String.format(DEBUGFORMAT,event.getClass(), event.getJsonPayload()));
+
+        if (!variantsAnalysisRepository.existsById(VariantsAnalysis.fromJson(event.getJsonPayload()).getAnalysisId())) {
+            throw new EventEntityDoesNotExistException();
+        }
+
         try {
             VariantsAnalysis variantsAnalysis = gson.fromJson(event.getJsonPayload(), VariantsAnalysis.class);
             variantsAnalysisRepository.delete(variantsAnalysis);
@@ -66,6 +78,11 @@ public class VariantEventListener {
     public void requirementCreated(RequirementCreatedEvent requirementCreatedEvent){
         logger.info("requirement created event");
         if(logger.isDebugEnabled())logger.debug(String.format(DEBUGFORMAT,requirementCreatedEvent.getClass(), requirementCreatedEvent.getJsonPayload()));
+
+        if (variantRequirementsRepository.existsById(VariantsRequirements.fromJson(requirementCreatedEvent.getJsonPayload()).getRequirementId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
+
         try {
             VariantsRequirements variantsRequirements = gson.fromJson(requirementCreatedEvent.getJsonPayload(), VariantsRequirements.class);
             variantRequirementsRepository.save(variantsRequirements);
@@ -80,6 +97,11 @@ public class VariantEventListener {
     public void requirementUpdated(RequirementUpdatedEvent requirementUpdatedEvent){
         logger.info("requirement updated event");
         if(logger.isDebugEnabled())logger.debug(String.format(DEBUGFORMAT,requirementUpdatedEvent.getClass(), requirementUpdatedEvent.getJsonPayload()));
+
+        if (!variantRequirementsRepository.existsById(VariantsRequirements.fromJson(requirementUpdatedEvent.getJsonPayload()).getRequirementId())) {
+            throw new EventEntityDoesNotExistException();
+        }
+
         try {
             VariantsRequirements variantsRequirements = gson.fromJson(requirementUpdatedEvent.getJsonPayload(), VariantsRequirements.class);
             variantRequirementsRepository.save(variantsRequirements);
@@ -94,6 +116,11 @@ public class VariantEventListener {
     public void requirementDeleted(RequirementDeletedEvent event){
         logger.info("requirement deleted event");
         if(logger.isDebugEnabled())logger.debug(String.format(DEBUGFORMAT,event.getClass(), event.getJsonPayload()));
+
+        if (!variantRequirementsRepository.existsById(VariantsRequirements.fromJson(event.getJsonPayload()).getRequirementId())) {
+            throw new EventEntityDoesNotExistException();
+        }
+
         try {
             VariantsRequirements variantsRequirements = gson.fromJson(event.getJsonPayload(), VariantsRequirements.class);
             variantRequirementsRepository.delete(variantsRequirements);
